b/174835544 Refactor build to support binary signing (#322)
* Move steps to create symbol package into makefile
* Add target to build installer snapshot
* Add target to build and sign installer from installer snapshot

+++ sources/scripts/clean.ps1
@@ -31,8 +31,16 @@ Resolve-Path -Path "$SourcesRoot\Google.Solutions.*\bin" |
 Resolve-Path -Path "$SourcesRoot\Google.Solutions.*\obj" | 
 	% { Remove-Item -Recurse -Force $_ }
 	
-# Delete obj directories
-Remove-Item -Recurse -Force "$SourcesRoot\installer\bin"
+# Delete installer directories
+if (Test-Path "$SourcesRoot\installer\obj") {
+	Remove-Item -Recurse -Force "$SourcesRoot\installer\obj"
+}
+
+if (Test-Path "$SourcesRoot\installer\bin") {
+	Remove-Item -Recurse -Force "$SourcesRoot\installer\bin"
+}
 
 # Delete packages
-Remove-Item -Recurse -Force "$SourcesRoot\packages"
+if (Test-Path "$SourcesRoot\packages") {
+	Remove-Item -Recurse -Force "$SourcesRoot\packages"
+}
